report, update history medical, enter medical

+++ database/migrations/2021_02_19_030132_create_table_medical_examinations_table.php
@@ -17,7 +17,7 @@ class CreateTableMedicalExaminationsTable extends Migration
             $table->increments('id');
             $table->date('date_examination')->nullable();
             $table->integer('circuit')->nullable();
-            $table->integer('temperature')->nullable();
+            $table->float('temperature')->nullable();
             $table->integer('breathing')->nullable();
             $table->integer('blood_pressure')->nullable();
             $table->string('diagnostic')->nullable();
